+++ pkg/rule/builtin/builtin.go
@@ -5,4 +5,5 @@ import "github.com/openserbia/doclint/pkg/rule"
 // Register adds every built-in rule to reg.
 func Register(reg *rule.Registry) {
 	reg.Register(DetailsBlankLine{})
+	reg.Register(TableColumnCount{})
 }